Improve precision of points when building `AreaTree`
By reading `double`s instead of `float` from the `PathIterator`, `AreaMeta` is better able to calculate angles to
determine whether a subpath is a hole or not. This is especially crucial at angles near +/-180 degrees.

+++ src/main/java/net/rptools/maptool/client/ui/zone/vbl/AreaMeta.java
@@ -103,7 +103,7 @@ public class AreaMeta {
     return isHole;
   }
 
-  public void addPoint(float x, float y) {
+  public void addPoint(double x, double y) {
     // Cut out redundant points
     // TODO: This works ... in concept, but in practice it can create holes that pop outside of
     // their parent bounds
